entrypoint/requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.0
pandas>=2.2
numpy>=1.26
tqdm>=4.66
geopy>=2.4
